Refinement of Lotes Page for Improved Functionality and User Experience
Updated the terminology throughout the application, changing all references of "Contenedor" to "Bandeja" to align with user expectations and provide a more accurate representation.
Added unique default names for each new "Bandeja" created during the addition of a new product, ensuring easy identification as "Bandeja A", "Bandeja B", etc.
Adjusted the + button for creating a new product, making it larger, more prominent, and providing a fully rounded design for better visual appeal.
Ensured that newly created products are immediately displayed in the sidebar list upon creation, improving the seamless flow of adding and selecting products.
Enhanced the title in the "Mediciones de Secado" section to make it bold and display the product name instead of the product code for a clearer and more user-friendly interface.